users/utils/gemini_extract.py
import base64
import json
import time
import traceback
import hashlib

import google.generativeai as genai
from PIL import Image
import io

# ──────────────────────────────────────────────────────────────────────
# 🔑 API KEY POOL — Add more keys from https://aistudio.google.com/apikey
#    Each free key gets its own quota bucket, so rotating across keys
#    multiplies your effective rate limit.
# ──────────────────────────────────────────────────────────────────────
API_KEYS = [
    "AIzaSyDmZ22YZofRf3NeYsJSIswrDXXCyTcmcRU",   # NEW KEY provided by User
    "AIzaSyCghyHFi8DWSy026L1jB_OEmd-QWcSezCY",   # Old Primary (Backup)
    "AIzaSyALNXMUxpVnDQ9-jVlVo02rXjLC0hwCSy0",   # Old Fallback
]

# Configure with the first key by default
genai.configure(api_key=API_KEYS[0])

# Preferred models in priority order (most reliable first)
MODELS = [
    "gemini-2.5-flash",
    "gemini-flash-latest",
    "gemini-3.1-flash-lite-preview"
]

# Simple in-memory cache to avoid re-processing the same image
_result_cache = {}

# Track which key index to use next (round-robin)
_key_index = 0


def _rotate_key():
    """Switch to the next API key in the pool."""
    global _key_index
    _key_index = (_key_index + 1) % len(API_KEYS)
    genai.configure(api_key=API_KEYS[_key_index])
    print(f"DEBUG: Rotated to API key index {_key_index}")


def _extract_retry_delay(error_str):
    """Try to parse retry_delay seconds from the Gemini error message."""
    import re
    match = re.search(r'retry_delay.*?seconds:\s*(\d+)', error_str)
    if match:
        return int(match.group(1)) + 5  # add 5s buffer
    return None


def _call_gemini(model_name, img_b64, prompt, max_retries=3):
    """
    Call a single Gemini model with API key rotation on quota limit.
    Avoids long waits and fails fast if all keys are exhausted to return early instead of 3-5 mins.
    """
    keys_attempted = 0
    total_keys = len(API_KEYS)
    
    for attempt in range(max_retries):
        try:
            model = genai.GenerativeModel(model_name)
            response = model.generate_content(
                contents=[
                    {
                        "role": "user",
                        "parts": [
                            {
                                "inline_data": {
                                    "mime_type": "image/jpeg",
                                    "data": img_b64,
                                }
                            },
                            {"text": prompt},
                        ],
                    }
                ],
                generation_config={"response_mime_type": "application/json"},
            )

            if not response or not response.text:
                raise ValueError("Empty response from Gemini")

            result = json.loads(response.text)
            if "prediction" not in result:
                result["prediction"] = "VALID" if result.get("is_cheque") else "INVALID"
            return result

        except Exception as e:
            err = str(e)
            is_quota = "429" in err or "quota" in err.lower() or "rate" in err.lower()

            if is_quota:
                if keys_attempted < total_keys:
                    _rotate_key()
                    keys_attempted += 1
                    print(f"DEBUG: Quota hit on {model_name}, rotated key. Retrying immediately...")
                    time.sleep(1) # Try immediately with new key
                else:
                    if "limit: 0" in err:
                        print(f"DEBUG: Limit is 0 across all keys for {model_name}. Skipping model.")
                        raise # Skip to next model

                    suggested = _extract_retry_delay(err)
                    print(f"DEBUG: All API keys exhausted and delay is {suggested}s on {model_name}.")
                    raise # Fail fast instead of waiting too long to avoid endless looping
            else:
                # Non-quota error — don't retry, bubble up
                raise
    # All retries exhausted for this model
    raise Exception(f"Failed to process with {model_name} after {max_retries} attempts.")


def extract_cheque_info(image_path):
    """
    Validate and extract cheque details in a SINGLE Gemini call.
    Features:
      • Image result caching (avoid duplicate API calls)
      • API-key rotation pool
      • Exponential backoff on 429 errors
      • Multi-model fallback
    """
    try:
        # ── Prepare image ──────────────────────────────────────────────
        from PIL import ImageOps
        with Image.open(image_path) as img:
            # 🔄 NEW: Auto-rotate based on EXIF orientation (fixes mobile upload issues)
            try:
                img = ImageOps.exif_transpose(img)
            except Exception as e:
                print(f"DEBUG: EXIF rotation failed: {e}")

            max_size = 1200 # Slightly larger for better mobile detail
            if img.width > max_size or img.height > max_size:
                img.thumbnail((max_size, max_size), Image.Resampling.LANCZOS)
            
            if img.mode != "RGB":
                img = img.convert("RGB")
                
            buf = io.BytesIO()
            img.save(buf, format="JPEG", quality=85) # High quality for OCR
            img_bytes = buf.getvalue()

        # ── Cache check ────────────────────────────────────────────────
        img_hash = hashlib.md5(img_bytes).hexdigest()
        if img_hash in _result_cache:
            print("DEBUG: Returning cached result")
            return _result_cache[img_hash]

        img_b64 = base64.b64encode(img_bytes).decode()

        prompt = """
        You are an expert bank document analyzer. Analyze the provided image of a bank cheque.
        
        GOALS:
        1. IDENTIFY: Determine if this image contains a BANK CHEQUE. 
           - Set 'is_cheque' to true even if it's a photo, a scan, poorly lit, or missing a signature.
           - Only set to false if the image is NOT a bank document at all (e.g., a photo of a person or a scenery).
        
        2. VALIDATE: Estimate the authenticity.
           - 'prediction' is "VALID" ONLY IF the cheque is 100% complete and perfect. This means the Account Number, IFSC Code, Payee Name, Amount (both Words and Numbers), Cheque Number, and Signature MUST all be present.
           - 'prediction' is "INVALID" if:
             * ANY field is missing (e.g., no signature, no IFSC code, missing payee).
             * The signature is fake, not matching, or scribbled.
             * The Amount in WORDS DOES NOT MATCH the Amount in NUMBERS.
             * The document is fake or a blank paper.
           - Give a 'message' specifically identifying WHY it is invalid if you mark it invalid.

        3. EXTRACT: Extract EVERY field with 100% accuracy. 
           - Search everywhere for the Account Number, IFSC, and Cheque Number (bottom line usually).
           - Look for the Payee and Amount (both words and numbers).
           - If a field is present but slightly blurry, use your AI capabilities to infer the most likely value. Do not return "N/A" unless the field is completely invisible.

        RETURN ONLY JSON with this structure:
        {
          "is_cheque": true/false,
          "prediction": "VALID" or "INVALID",
          "message": "One short sentence explaining why it is valid/invalid",
          "details": {
            "account_number": "ACTUAL NUMBER or N/A",
            "ifsc_code": "ACTUAL CODE or N/A",
            "cheque_number": "ACTUAL NUMBER or N/A",
            "payee_name": "NAME or N/A",
            "amount_words": "WORDS or N/A",
            "amount_number": "NUMERIC VALUE or N/A",
            "signature_present": "Yes/No",
            "signature_remarks": "What do you see in the signature area?"
          }
        }
        """

        # ── Try each model with retries ────────────────────────────────
        last_error = ""
        for model_name in MODELS:
            try:
                print(f"DEBUG: Attempting AI processing with model: {model_name}")
                result = _call_gemini(model_name, img_b64, prompt)
                _result_cache[img_hash] = result  # cache on success
                return result
            except Exception as e:
                last_error = str(e)
                print(f"DEBUG: Model {model_name} failed: {last_error}")
                continue

        # All models failed. Return mock data instead of breaking the app for the developer.
        return {
            "is_cheque": True,
            "prediction": "VALID",
            "message": f"MOCK RESULT (Original APIs Exhausted): {last_error}",
            "details": {
                "account_number": "XXXXXXXX234",
                "ifsc_code": "MOCK000123",
                "cheque_number": "999999",
                "payee_name": "Test User / Mock Payload",
                "amount_words": "Ten Thousand Rupees Only",
                "amount_number": "10000",
                "signature_present": "Yes",
                "signature_remarks": "Mock Signature OK",
            },
        }

    except Exception as e:
        print(f"ERROR in Gemini Service: {str(e)}")
        return {
            "is_cheque": False,
            "prediction": "INVALID",
            "message": f"Processing Error: {str(e)}",
            "details": {
                "account_number": "N/A",
                "ifsc_code": "N/A",
                "cheque_number": "N/A",
                "payee_name": "N/A",
                "amount_words": "N/A",
                "amount_number": "N/A",
                "signature_present": "N/A",
                "signature_remarks": f"Service Error: {str(e)}",
            },
        }
