requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.1+cpu
torchvision==0.15.2+cpu
google-generativeai
django
pillow
opencv-python-headless
numpy
joblib
scikit-learn
matplotlib
gunicorn
whitenoise
